backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "authlib>=1.6.9",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "openai>=2.28.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
    "websockets>=16.0",
]
